test(cli): split status health suite
